+++ src/Http/Resources/ProviderStatusResource.php
@@ -41,8 +41,8 @@ class ProviderStatusResource extends JsonResource
         return [
             'provider' => (new ProviderResource($this->provider))->resolve(),
             'active_calls_count' => $this->activeCallsCount(),
-            'online_status' => $this->provider->onlineStatus(),
-            'online_status_verbose' => Definitions::OnlineStatus[$this->provider->onlineStatus()],
+            'online_status' => $this->provider->getProviderOnlineStatus(),
+            'online_status_verbose' => Definitions::OnlineStatus[$this->provider->getProviderOnlineStatus()],
             'unread_threads_count' => $this->unreadThreadsCount(),
             'pending_friends_count' => $this->pendingFriendsCount(),
             'settings' => Messenger::getProviderMessenger($this->provider)->toArray(),