+++ tests/Fixtures/CompanyModel.php
@@ -19,12 +19,12 @@ class CompanyModel extends User implements MessengerProvider, Searchable
 
     protected $guarded = [];
 
-    public function name(): string
+    public function getProviderName(): string
     {
         return strip_tags(ucwords($this->company_name));
     }
 
-    public function getAvatarColumn(): string
+    public function getProviderAvatarColumn(): string
     {
         return 'avatar';
     }